Rename ez keymaps for userspace
