--- keyboards/ergodox_ez/keymaps/drashna-custom/rules.mk
@@ -1,12 +1,0 @@
-TAP_DANCE_ENABLE  = yes
-SLEEP_LED_ENABLE  = no  # Breathing sleep LED during USB suspend
-COMMAND_ENABLE    = yes  # Commands for debug and configuration
-RGBLIGHT_ENABLE   = yes
-MIDI_ENABLE       = no
-CONSOLE_ENABLE    = no
-
-EXTRAFLAGS        = -flto
-
-ifndef QUANTUM_DIR
-	include ../../../../Makefile
-endif
