+++ keyboards/ergodox_ez/keymaps/drashna-ez/rules.mk
@@ -1,0 +1,6 @@
+RGBLIGHT_ENABLE = yes
+RGBLIGHT_ANIMATION = no
+
+ifndef QUANTUM_DIR
+	include ../../../../Makefile
+endif
