+++ keyboards/ergodox_ez/keymaps/drashna/config.h
@@ -1,11 +1,19 @@
-
+#ifndef CONFIG_USER_H
+#define CONFIG_USER_H
 #include "../../config.h"
 
 
+#define TAPPING_TERM 200
 
-// make this easy to toggle behavior, so that it can be more easily toggled
-#define LAYER_UNDERGLOW_LIGHTING  
+#ifdef RGBLIGHT_ENABLE
+#undef RGBLIGHT_SAT_STEP 
+#define RGBLIGHT_SAT_STEP 12
+#define RGBLIGHT_EFFECT_KNIGHT_LENGTH 7
+#define RGBLIGHT_EFFECT_SNAKE_LENGTH 7
+#define RGBLIGHT_EFFECT_BREATHE_CENTER 1
+#endif // RGBLIGHT_ENABLE
 
-#ifdef LAYER_UNDERGLOW_LIGHTING
-#undef RGBLIGHT_ANIMATIONS
+#undef DEBOUNCE
+#define DEBOUNCE    5
+
 #endif
