+++ keyboards/ergodox_ez/keymaps/drashna/rules.mk
@@ -1,5 +1,11 @@
-RGBLIGHT_ENABLE = yes
-RGBLIGHT_ANIMATION = no
+TAP_DANCE_ENABLE  = yes
+SLEEP_LED_ENABLE  = no  # Breathing sleep LED during USB suspend
+COMMAND_ENABLE    = yes  # Commands for debug and configuration
+RGBLIGHT_ENABLE   = yes
+MIDI_ENABLE       = no
+CONSOLE_ENABLE    = no
+
+EXTRAFLAGS        = -flto
 
 ifndef QUANTUM_DIR
 	include ../../../../Makefile
