+++ users/drashna/rules.mk
@@ -1,0 +1,1 @@
+SRC += drashna.c
